refactor: extract SubscriptionLifecycleController from users
UsersControllers fused auth, five OAuth providers, and subscription
lifecycle in one 1,542-line file. Carve the cancel/pause/resume/status/
feedback block into a focused SubscriptionLifecycleController with its
own constructor-injected deps (UsersService + Knex), wired once in the
UserRouter composition root. Pure move: every route path, method,
middleware chain, status code, and response body is unchanged; the moved
tests move verbatim (only the class name and the dropped unused
authService arg differ). No new data_layer import edges — pnpm arch
layer-skip count holds at 255. First of the #4201 phases.
